Revert "Backup Support for SQL Server in Azure VM"

+++ src/RecoveryServices/RecoveryServices.Backup.Providers/Providers/AzureFilesPsBackupProvider.cs
@@ -17,6 +17,7 @@ using Microsoft.Azure.Commands.RecoveryServices.Backup.Cmdlets.ServiceClientAdap
 using Microsoft.Azure.Commands.RecoveryServices.Backup.Helpers;
 using Microsoft.Azure.Commands.RecoveryServices.Backup.Properties;
 using Microsoft.Azure.Management.Internal.Resources.Models;
+using Microsoft.Azure.Management.Internal.Resources.Utilities.Models;
 using Microsoft.Azure.Management.RecoveryServices.Backup.Models;
 using Microsoft.Rest.Azure.OData;
 using System;
@@ -26,6 +27,7 @@ using System.Management.Automation;
 using CmdletModel = Microsoft.Azure.Commands.RecoveryServices.Backup.Cmdlets.Models;
 using RestAzureNS = Microsoft.Rest.Azure;
 using ServiceClientModel = Microsoft.Azure.Management.RecoveryServices.Backup.Models;
+using SystemNet = System.Net;
 
 namespace Microsoft.Azure.Commands.RecoveryServices.Backup.Cmdlets.ProviderModel
 {
@@ -83,38 +85,29 @@ namespace Microsoft.Azure.Commands.RecoveryServices.Backup.Cmdlets.ProviderModel
 
             AzureFileShareItem item = (AzureFileShareItem)ProviderData[ItemParams.Item];
 
-            AzureFileshareProtectedItem properties = new AzureFileshareProtectedItem();
-
-            return EnableOrModifyProtection(disableWithRetentionData: true);
-
-        }
-
-        public RestAzureNS.AzureOperationResponse DisableProtectionWithDeleteData()
-        {
-            string vaultName = (string)ProviderData[VaultParams.VaultName];
-            string vaultResourceGroupName = (string)ProviderData[VaultParams.ResourceGroupName];
-            bool deleteBackupData = ProviderData.ContainsKey(ItemParams.DeleteBackupData) ?
-                (bool)ProviderData[ItemParams.DeleteBackupData] : false;
-
-            ItemBase itemBase = (ItemBase)ProviderData[ItemParams.Item];
-
-            AzureFileShareItem item = (AzureFileShareItem)ProviderData[ItemParams.Item];
-
             string containerUri = "";
             string protectedItemUri = "";
             AzureFileshareProtectedItem properties = new AzureFileshareProtectedItem();
 
-            ValidateAzureFileShareDisableProtectionRequest(itemBase);
+            if (deleteBackupData)
+            {
+                //Disable protection and delete backup data
+                ValidateAzureFileShareDisableProtectionRequest(itemBase);
 
-            Dictionary<UriEnums, string> keyValueDict = HelperUtils.ParseUri(item.Id);
-            containerUri = HelperUtils.GetContainerUri(keyValueDict, item.Id);
-            protectedItemUri = HelperUtils.GetProtectedItemUri(keyValueDict, item.Id);
+                Dictionary<UriEnums, string> keyValueDict = HelperUtils.ParseUri(item.Id);
+                containerUri = HelperUtils.GetContainerUri(keyValueDict, item.Id);
+                protectedItemUri = HelperUtils.GetProtectedItemUri(keyValueDict, item.Id);
 
-            return ServiceClientAdapter.DeleteProtectedItem(
-                                containerUri,
-                                protectedItemUri,
-                                vaultName: vaultName,
-                                resourceGroupName: vaultResourceGroupName);
+                return ServiceClientAdapter.DeleteProtectedItem(
+                                    containerUri,
+                                    protectedItemUri,
+                                    vaultName: vaultName,
+                                    resourceGroupName: vaultResourceGroupName);
+            }
+            else
+            {
+                return EnableOrModifyProtection(disableWithRetentionData: true);
+            }
         }
 
         public List<ContainerBase> ListProtectionContainers()
@@ -483,8 +476,7 @@ namespace Microsoft.Azure.Commands.RecoveryServices.Backup.Cmdlets.ProviderModel
             }
 
             //inquiry
-            AzureWorkloadProviderHelper.TriggerInquiry(vaultName, vaultResourceGroupName,
-                storageContainerName, ServiceClientModel.WorkloadType.AzureFileShare);
+            TriggerInquiry(vaultName, vaultResourceGroupName, storageContainerName);
 
             //get protectable item
             WorkloadProtectableItemResource protectableObjectResource = null;
@@ -538,6 +530,37 @@ namespace Microsoft.Azure.Commands.RecoveryServices.Backup.Cmdlets.ProviderModel
                 }
             }
             return protectableObjectResource;
+        }
+
+        private void TriggerInquiry(string vaultName, string vaultResourceGroupName,
+            string storageContainerName)
+        {
+            ODataQuery<BMSContainersInquiryQueryObject> queryParams = new ODataQuery<BMSContainersInquiryQueryObject>(
+                q => q.WorkloadType
+                     == ServiceClientModel.WorkloadType.AzureFileShare);
+            string errorMessage = string.Empty;
+            var inquiryResponse = ServiceClientAdapter.InquireContainer(
+               storageContainerName,
+               queryParams,
+               vaultName,
+               vaultResourceGroupName);
+
+            var operationStatus = TrackingHelpers.GetOperationResult(
+                inquiryResponse,
+                operationId =>
+                    ServiceClientAdapter.GetContainerRefreshOrInquiryOperationResult(
+                        operationId,
+                        vaultName: vaultName,
+                        resourceGroupName: vaultResourceGroupName));
+
+            //Now wait for the operation to Complete
+            if (inquiryResponse.Response.StatusCode
+                    != SystemNet.HttpStatusCode.NoContent)
+            {
+                errorMessage = string.Format(Resources.TriggerEnquiryFailureErrorCode,
+                    inquiryResponse.Response.StatusCode);
+                Logger.Instance.WriteDebug(errorMessage);
+            }
         }
 
         private List<ContainerBase> GetRegisteredStorageAccounts(string vaultName = null,
@@ -675,11 +698,6 @@ namespace Microsoft.Azure.Commands.RecoveryServices.Backup.Cmdlets.ProviderModel
             }
 
             return itemModels;
-        }
-
-        public void RegisterContainer()
-        {
-            throw new NotImplementedException();
         }
 
         private RestAzureNS.AzureOperationResponse EnableOrModifyProtection(bool disableWithRetentionData = false)
@@ -895,11 +913,6 @@ namespace Microsoft.Azure.Commands.RecoveryServices.Backup.Cmdlets.ProviderModel
             {
                 throw new ArgumentException(string.Format(Resources.AzureFileTargetSANameMissingException));
             }
-        }
-
-        public List<PointInTimeBase> GetLogChains()
-        {
-            throw new NotImplementedException();
         }
     }
 }
